requirements.txt
Flask
pandas
numpy
requests
scikit-learn
joblib
gunicorn
python-dotenv